# api/openapi.yaml
openapi: 3.0.3
info:
  title: 自动化测试管理平台 API
  description: |
    面向服务器 POC 团队的一站式性能测试平台 API 文档。
    
    ## 功能模块
    - 服务器管理：服务器信息录入、分组、状态监控
    - 脚本中心：测试脚本上传、版本管理、参数配置、输出解析
    - 测试任务：任务创建、调度、执行、监控
    - 测试结果：结果解析、判定、对比、趋势分析
    - 报告中心：报告生成、查看、导出
    
    ## 认证方式
    当前版本暂无用户认证，后续版本将支持 Token 认证。
    
  version: 1.0.2
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: http://test.example.com/api/v1
    description: 测试环境
  - url: http://prod.example.com/api/v1
    description: 生产环境
...
paths:
...
  /server-groups:
    get:
      tags:
        - server-groups
      summary: 获取分组列表
      operationId: listServerGroups
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServerGroup'
    
    post:
      tags:
        - server-groups
      summary: 创建分组
      operationId: createServerGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerGroupCreate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/ServerGroup'
  /server-groups/{id}:
    put:
      tags:
        - server-groups
      summary: 更新分组
      operationId: updateServerGroup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerGroupCreate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    delete:
      tags:
        - server-groups
      summary: 删除分组
      operationId: deleteServerGroup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================
  # 服务器管理
  # ============================================================
  /servers:
    get:
      tags:
        - servers
      summary: 获取服务器列表
      operationId: listServers
      parameters:
        - name: name
          in: query
          description: 服务器名称（模糊搜索）
          schema:
            type: string
        - name: status
          in: query
          description: 状态筛选
          schema:
            type: string
            enum: [online, offline, maintenance]
        - name: group_id
          in: query
          description: 分组ID
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      size:
                        type: integer
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Server'
    
    post:
      tags:
        - servers
      summary: 添加服务器
      operationId: createServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Server'
  /servers/{id}:
    get:
      tags:
        - servers
      summary: 获取服务器详情
      operationId: getServer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Server'
    
    put:
      tags:
        - servers
      summary: 更新服务器
      operationId: updateServer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    delete:
      tags:
        - servers
      summary: 删除服务器
      operationId: deleteServer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /servers/{id}/test:
    post:
      tags:
        - servers
      summary: 测试连接
      operationId: testServerConnection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: 连接成功
                  data:
                    type: object
                    properties:
                      connected:
                        type: boolean
                      latency:
                        type: integer
                        description: 连接延迟(ms)
  /servers/{id}/status:
    get:
      tags:
        - servers
      summary: 获取服务器状态
      operationId: getServerStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [online, offline, maintenance]
                      lastCheckAt:
                        type: string
                        format: date-time
                      cpuUsage:
                        type: number
                      memoryUsage:
                        type: number
  /servers/{id}/refresh:
    post:
      tags:
        - servers
      summary: 刷新服务器信息
      description: 自动识别CPU、内存、磁盘、操作系统等信息
      operationId: refreshServerInfo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Server'
...
  /scripts/{id}:
    get:
      tags:
        - scripts
      summary: 获取脚本详情
      operationId: getScript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Script'
    
    put:
      tags:
        - scripts
      summary: 更新脚本
      operationId: updateScript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptUpdate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    delete:
      tags:
        - scripts
      summary: 删除脚本
      operationId: deleteScript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object

    # 服务器分组
    ServerGroup:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServerGroupCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string

    # 服务器
    Server:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        host:
          type: string
        port:
          type: integer
          default: 22
        username:
          type: string
        authType:
          type: string
          enum: [password, ssh_key]
        osType:
          type: string
        osVersion:
          type: string
        cpuCores:
          type: integer
        cpuModel:
          type: string
        memorySize:
          type: string
        memoryTotalMb:
          type: integer
        diskInfo:
          type: object
          additionalProperties: true
        groupId:
          type: integer
          format: int64
        groupName:
          type: string
        tags:
          type: array
          items:
            type: string
        remark:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance]
        lastCheckAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServerCreate:
      type: object
      required:
        - name
        - host
        - port
        - username
        - authType
        - authSecret
      properties:
        name:
          type: string
          maxLength: 100
        host:
          type: string
          maxLength: 255
        port:
          type: integer
          default: 22
        username:
          type: string
          maxLength: 50
        authType:
          type: string
          enum: [password, ssh_key]
        authSecret:
          type: string
          description: 密码或SSH私钥
        groupId:
          type: integer
          format: int64
        tags:
          type: array
          items:
            type: string
        remark:
          type: string

    # 脚本
    Script:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        scriptType:
          type: string
          enum: [shell, python]
        testCategory:
          type: string
        description:
          type: string
        currentVersion:
          type: string
        lifecycleMode:
          type: string
          enum: [simple, full]
        hasDeploy:
          type: boolean
        hasCleanup:
          type: boolean
        deployEntry:
          type: string
        cleanupEntry:
          type: string
        entryFile:
          type: string
        fileList:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              size:
                type: integer
              isEntry:
                type: boolean
        parameters:
          type: object
          additionalProperties: true
        parseRules:
          type: object
          additionalProperties: true
        defaultTimeout:
          type: integer
        defaultRetry:
          type: integer
        isBuiltin:
          type: boolean
        status:
          type: string
          enum: [enabled, disabled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScriptUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
          additionalProperties: true
        parseRules:
          type: object
          additionalProperties: true
        defaultTimeout:
          type: integer
        defaultRetry:
          type: integer
        status:
          type: string
          enum: [enabled, disabled]
...